Separate notifications from providers
