+++ packages/neon-api/src/plugin.ts
@@ -1,7 +1,7 @@
 import * as neoCli from "./provider/neoCli";
 import * as neonDB from "./provider/neonDB";
 import * as neoscan from "./provider/neoscan";
-import * as notifications from "./provider/notifications";
+import * as notifications from "./notifications";
 
 export { neoCli, neoscan, neonDB, notifications };
 export * from "./funcs";
